Wire AgenticLoop into Hilt via AiAssistantModule
