+++ libs/ai-assistant/feature/src/main/kotlin/com/woocommerce/android/aiassistant/di/AiAssistantModule.kt
@@ -1,5 +1,12 @@
 package com.woocommerce.android.aiassistant.di
 
+import com.woocommerce.android.aiassistant.core.chat.ChatService
+import com.woocommerce.android.aiassistant.core.chat.ToolRegistry
+import com.woocommerce.android.aiassistant.core.loop.AgenticLoop
+import com.woocommerce.android.aiassistant.core.loop.AgenticLoopImpl
+import com.woocommerce.android.aiassistant.core.loop.ConservativeRetryPolicy
+import com.woocommerce.android.aiassistant.core.loop.NoOpToolRegistry
+import com.woocommerce.android.aiassistant.core.loop.RetryPolicy
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -17,4 +24,21 @@ internal object AiAssistantModule {
         ignoreUnknownKeys = true
         encodeDefaults = false
     }
+
+    @Provides
+    @Singleton
+    fun provideAgenticLoop(
+        chatService: ChatService,
+        toolRegistry: ToolRegistry,
+        retryPolicy: RetryPolicy,
+        @AiAssistantJson json: Json,
+    ): AgenticLoop = AgenticLoopImpl(chatService, toolRegistry, retryPolicy, json)
+
+    @Provides
+    @Singleton
+    fun provideToolRegistry(): ToolRegistry = NoOpToolRegistry()
+
+    @Provides
+    @Singleton
+    fun provideRetryPolicy(): RetryPolicy = ConservativeRetryPolicy
 }
